changed Quesstions page, added some functions for adaptive design
