+++ Tracker Hub/FrontEnd/Adaptive design/StringData.swift
